% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aisdk.bioc-package.R
\docType{package}
\name{aisdk.bioc-package}
\alias{aisdk.bioc}
\alias{aisdk.bioc-package}
\title{aisdk.bioc: Bioconductor semantic extensions for aisdk}
\description{
Extension package that hosts Bioconductor-specific semantic adapters and
workflow hints for the \code{aisdk} runtime.
}
\author{
\strong{Maintainer}: Yonghe Xia \email{xiayh17@gmail.com}

}
\keyword{internal}
